test(daemon): split shutdown coverage below size ceiling
